remove bool and do c99
compatability is ass on debian trash

+++ src/motor_test.c
@@ -24,7 +24,7 @@ int main(int argc, const char* argv[])
     cfg.picture_size = (float)atof(argv[3]);
     cfg.steps_per_revolution_yaw = (size_t)atoi(argv[4]);
     cfg.steps_per_revolution_pitch = (size_t)atoi(argv[5]);
-    cfg.motor_speed = (size_t)atol(argv[6]);
+    cfg.motor_speed = (size_t)atoi(argv[6]);
     set_config(&cfg);
 
     // for(int i = 0; i < NUM_POINTS; i++) {
